Added ir-compression option.
